extract commit analyser code
